requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.35
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
